refactor: 외부로 노출되는 /api/trpc를 /api로 변경
